DateFormat: fix format "dd.MM.y"
The term "y" is handled together with "yy" and "yyy" (but not "yyyy").
This lead to actual valid dates with 4-digit years not being accepted
anymore, because the corresponding "parseRegExp" only allowed 3
digits. When we extend this to 4 digits, we have to adjust some other
parts as well. The existing implementation was re-implemented according
to the rules specified by Java's SimpleDateFormat. If a digit year is
entered, we now assume a year between -80/+20 years from the start date
(instead of -50/+50).

390261 Subentitäts Datum-Attribut wird in Maske immer rot beim setzen