TileGridLayout: mark tiles as 'animating' while changing bounds
Add a promise to the animated container while tile bounds are changed
with a jQuery animation. This allows the HtmlComponent to wait for the
animation to finish before validating the content (otherwise, wrong
dimensions might be measured). It works similarly to the animate-*
classes that are used for CSS animations, which we cannot use in this
case, because no animationEnd would is fired after the jQuery animation
has been completed.

390356